data-processor/requirements.txt
flask
psycopg2-binary
paho-mqtt